Remove redundant code

+++ sim_demo_collector/util/pcd_util.py
@@ -27,7 +27,7 @@ def depth2pcd(
     z = depth
     x = (u - cx) / fx * z
     y = (v - cy) / fy * z
-    pcd = np.stack([x, y, z], axis=-1, dtype=np.float32)
+    pcd = np.stack([x, y, z], axis=-1)
     # Apply bounding box mask
     bb_mask = np.ones_like(z, dtype=bool)
     if bounding_box is not None:
